ATUALIZAÇÃO
FIXES A CODE

+++ TaskManagerApp-Backend/TaskManagerAPI/Repositories/TaskRepository.cs
@@ -16,35 +16,45 @@ namespace TaskManagerAPI.Repositories
             _context = context;
         }
 
-        public async Task<IEnumerable<Task>> GetTasks()
+        public async Task<IEnumerable<Models.Task>> GetTasks()
         {
             return await _context.Tasks.ToListAsync();
         }
 
-        public async Task<Task> GetTask(int id)
+        public async Task<Models.Task> GetTask(int id)
         {
             return await _context.Tasks.FindAsync(id);
         }
 
-        public async Task<Task> AddTask(Task task)
+        public async Task<Models.Task> AddTask(Models.Task task)
         {
             _context.Tasks.Add(task);
             await _context.SaveChangesAsync();
             return task;
         }
 
-        public async Task<Task> UpdateTask(Task task)
+        public async Task<bool> UpdateTask(Models.Task task)
         {
             _context.Entry(task).State = EntityState.Modified;
             await _context.SaveChangesAsync();
+            return true;
+        }
+
+        public async Task<Models.Task> DeleteTask(int id)
+        {
+            Models.Task task = await _context.Tasks.FindAsync(id);
+
+            if (task == null)
+            {
+                return null;
+            }
+
+            _context.Tasks.Remove(task);
+            await _context.SaveChangesAsync();
+
             return task;
         }
 
-        public async Task DeleteTask(int id)
-        {
-            var task = await _context.Tasks.FindAsync(id);
-            _context.Tasks.Remove(task);
-            await _context.SaveChangesAsync();
-        }
+
     }
 }
